Switched to new blocking. Removed old blocking code
Former-commit-id: 9032f16db3e1a8ae5613b2e958fb04607dafcd28

+++ silk2/core/src/main/scala/de/fuberlin/wiwiss/silk/linkspec/condition/Aggregation.scala
@@ -15,7 +15,7 @@ case class Aggregation(required : Boolean, weight : Int, operators : Traversable
     aggregator.evaluate(weightedValues).toList
   }
 
-  override def index(instance : Instance) : Set[Seq[Double]] =
+  override def index(instance : Instance) : Set[Seq[Int]] =
   {
     val indexSets = operators.map(_.index(instance))
     aggregator.aggregateIndexes(indexSets)
